der: add minimal GeneralStringRef type
